fix(beta.app): infinite loop error crashing app on re-render
